requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-appconf==1.0.3
django-extensions==2.2.5
django-filter==2.2.0
django-imagekit==4.0.2
djangorestframework==3.10.3
enum34==1.1.6
exif==0.8.1
idna==2.8
Markdown==3.1.1
pilkit==2.0
Pillow==6.2.1
python-resize-image==1.1.19
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
uWSGI==2.0.18
